Allow changing the container port
The default image uses port 80 so when first bringing up the service it must be set to port 80 (unless you change the default image). There is a good chance that the next image being deployed uses a different port, so it needs to be possible to change it over time using a secret value.